# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UniStore API",
        "description": "API REST UniStore - BIT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mes notifications",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "Liste récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/lu": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer comme lu",
                "operationId": "1f391cb8c9a91bf772ac7a1257f6c9d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marqué comme lu"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/tout-lu": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Tout marquer comme lu",
                "operationId": "08aeb6b4cf7968f82d254967e57f8f3f",
                "responses": {
                    "200": {
                        "description": "Toutes marquées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Supprimer une notification",
                "operationId": "54c4eb5ec5dc0d1e2916cec2fced1056",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commandes/{commande_id}/paiement": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Voir le paiement d'une commande",
                "operationId": "c66ed8f899e87116f2ff7c530eb80353",
                "parameters": [
                    {
                        "name": "commande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paiement trouvé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Payer une commande",
                "operationId": "4bb11dc75184fff72d495461f72350e7",
                "parameters": [
                    {
                        "name": "commande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "methode"
                                ],
                                "properties": {
                                    "methode": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "mobile_money",
                                            "carte_bancaire"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement initialisé"
                    },
                    "400": {
                        "description": "Statut invalide ou déjà payé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paiements/{id}/valider": {
            "put": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Valider paiement (Vendeur)",
                "operationId": "7a3800073ef5ca219e66d197e17565ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paiement validé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paiements/{id}/rembourser": {
            "put": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Rembourser paiement (Vendeur)",
                "operationId": "88a36462c6938b26e6a3725d1381bd1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remboursé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}